Refactor DnD into it's own component
